feat(kitchen-order): cancel kitchen order when kitchen staff cook

+++ src/app/features/customer/pages/order-cart/order-cart.component.ts
@@ -4,6 +4,9 @@ import { DrawerModule } from 'primeng/drawer';
 import { ButtonModule } from 'primeng/button';
 import { ContextMenu, ContextMenuModule } from 'primeng/contextmenu';
 import { BadgeModule } from 'primeng/badge'
+import { ConfirmDialogModule } from 'primeng/confirmdialog';
+
+import { ConfirmationService } from 'primeng/api';
 
 import { OrderItem } from '../../../../core/models/order-item.model';
 import { OrderItemCardComponent } from "../../components/order-item-card/order-item-card.component";
@@ -13,10 +16,12 @@ import { OrderService } from '../../../../core/services/order.service';
 @Component({
   selector: 'app-order-cart',
   standalone: true,
-  imports: [CommonModule, DrawerModule, ButtonModule, ContextMenuModule, BadgeModule, OrderItemCardComponent],
+  imports: [CommonModule, DrawerModule, ButtonModule, ContextMenuModule, BadgeModule, OrderItemCardComponent, ConfirmDialogModule],
+  providers: [ConfirmationService],
   templateUrl: './order-cart.component.html',
   styleUrl: './order-cart.component.css'
 })
+
 export class OrderCartComponent {
   @Input() visible!: boolean;
   @Output() visibleChange = new EventEmitter<boolean>();
@@ -27,6 +32,7 @@ export class OrderCartComponent {
 
   private cartService = inject(CartService);
   private orderService = inject(OrderService);
+  confirmationService = inject(ConfirmationService);
 
   isSubmitting = false;
 
@@ -63,20 +69,32 @@ export class OrderCartComponent {
   }
 
   confirmOrder() {
-    this.isSubmitting = true;
+    if (this.cartService.getCart().length === 0) return;
 
-    this.orderService.createOrder().subscribe({
-      next: data => {
-        localStorage.setItem('orderId', data['id']);
+    this.confirmationService.confirm({
+      header: 'Confirm Order',
+      message: 'Are you sure confirm order?',
+      icon: 'pi pi-check-circle',
+      acceptLabel: 'Confirm',
+      rejectLabel: 'Cancel',
+      acceptButtonStyleClass: 'p-button-info',
+      accept: () => {
+        this.isSubmitting = true;
 
-        const orderId = localStorage.getItem('orderId');
-
-        console.log(`Don hang ${orderId} da duoc tao!`);
-      },
-      error: err => console.log(err),
-      complete: () => {
-        console.log("Order thành công");
+        this.orderService.createOrder().subscribe({
+          next: data => {
+            localStorage.setItem('orderId', data['id']);
+    
+            const orderId = localStorage.getItem('orderId');
+    
+            console.log(`Don hang ${orderId} da duoc tao!`);
+          },
+          error: err => console.log(err),
+          complete: () => {
+            console.log("Order thành công");
+          }
+        });
       }
-    });
+    }); 
   }
 }
